requirements.txt
streamlit
numpy
Pillow
tensorflow==2.20.0
onnxruntime
tf2onnx
